+++ detectron/modeling/roi_heads/box_head/loss.py
@@ -59,7 +59,7 @@ class FastRCNNLossComputation(object):
             matched_idxs = matched_targets.get_field("matched_idxs")
 
             labels_per_image = matched_targets.get_field("labels")
-            labels_per_image = labels_per_image.int64()
+            labels_per_image = labels_per_image.int32()
 
             # Label background (below the low threshold)
             bg_inds = matched_idxs == Matcher.BELOW_LOW_THRESHOLD
